requirements.txt
# Python 3.11+
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
mutagen>=1.47.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0

# Desenvolvimento / testes
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
